Added simple code for making a get request and parsing the JSON data
This is a simple demonstration of the ESP32 board making a get request to an endpoint, receiving the JSON data, and printing the value of the specified JSON property. This will be used in the future for allowing the alarm to get information from the web server in order to update itself.